Optimized PublicOrleansTaskExtentions.Ignore() to avoid going through the async state machine unnecessarily.
I had this code in my repo for very long, and since this is a heavily used method, I think it's worth optimizing.

I did run some benchmarks, and the improvements in perf are from 2 times to 16 times (depending on whether the preceding task completed synchronously or not, and whether it failed or not). The point is, in all situations I found improvements in perf.

The Benchmark results are as follows:

#### Completed synchronously
```
                  Method |     AvrTime |    StdDev |          op/s |
------------------------ |------------ |---------- |-------------- |
        IgnoreUsingAwait | 137.0142 ns | 0.4176 ns |  7,298,515.37 |
 IgnoreUsingContinueWith |  14.4233 ns | 0.0593 ns | 69,332,185.35 |
```

#### Failed synchronously
```
                  Method |    AvrTime |    StdDev |       op/s |
------------------------ |----------- |---------- |----------- |
        IgnoreUsingAwait | 20.7958 us | 0.1675 us |  48,086.69 |
 IgnoreUsingContinueWith |  1.2535 us | 0.5995 us | 797,791.52 |
```
 
#### Completing asynchronously
```
                  Method |       AvrTime |      StdDev |         op/s |
------------------------ |-------------- |------------ |------------- |
        IgnoreUsingAwait | 1,163.1511 ns | 211.5803 ns |   859,734.17 |
 IgnoreUsingContinueWith |   668.6211 ns | 132.1161 ns | 1,495,615.61 |
``` 

#### Failing asynchronously
```
                  Method |    AvrTime |     StdDev |       op/s |
------------------------ |----------- |----------- |----------- |
        IgnoreUsingAwait | 26.9202 us | 12.1708 us |  37,146.83 |
 IgnoreUsingContinueWith |  4.3352 us |  0.0552 us | 230,672.58 |
```

The benchmark code itself is in this gist: https://gist.github.com/jdom/65cd644bbcab9a41a75a