@gabikliot 2015-12-10
Julian, is there any concern that task returned by CW (the `task.ContinueWith`) will be broken with CancellataionException in case that CW does not run (if the actual task succeeded), since you specified `TaskContinuationOptions.OnlyOnFaulted`. And in such a case the non-observed cancelled task will escalate to Finalizer thread with UnobservedException? Is this possible?