@gabikliot 2015-12-10
Did you use `TaskScheduler.Default` deliberately? Instead of `TaskScheduler.Current`.