@jdom 2015-12-10
@gabikliot there is no need to observe a canceled task (there is no raised, it is only surfaces as an OperationCanceledException if you await it, so no need to observe any of it, whether it is because the previous task was Canceled, or because the previous task ended as faulted so that the continuation task is Canceled).

Regarding the usage of `TaskScheduler.Default`, yes, it's deliberate. Nevertheless, because I used `TaskContinuationOptions.ExecuteSynchronously`, there is no background thread or anything, the TaskPoolScheduler guarantees that it can run in the same thread.