@gabikliot 2015-12-10
So a failed task results in Unobserved Exception if it is not awaited. 
But cancelled task does NOT result in Unobserved Exception if it is not awaited?

As for `TaskScheduler.Default` - I think based on what you wrote it even makes more sense to use TaskScheduler.Current. So we are running on some TaskScheduler X. The task we are about to ignore was created and runs on  TaskScheduler X. Now you are asking to do CW `ExecuteSynchronously` but on a different TaskScheduler? What is the benefit? Sounds like an extra overhead of switching to a different scheduler, instead of inlining and executing synchronously on X.
No? I got it wrong?