@gabikliot 2015-12-10
BTW, to be clear, I am not trying to pick up on that PR.
I think this PR is great!
Actually, I clearly remember that our initial version of `Ignore` used to do exactly that. Then I got second thoughts exactly about the questions I am asking you now, plus I saw a bunch of other places doing `Ignore` with `await` and not with `CW`, and so to be 1000% sure and err on the correctness side I switched it use `await`. 
I am totally fine to moving back to `CW`, just want to be 100% sure.