@jdom 2015-12-10
correct, a canceled task does not result in an exception. Period. Neither observed nor unobserved. It only materializes as an exception when awaiting it.

Regarding the scheduler, there is no need to *transition* to another thread or anything, the scheduler code just invokes the continuation without any further transitioning. `TaskScheduler.Current` *might* also do the same (not try to be smart about where to schedule it and avoid transitions), but it might not. I was told to do this by Stephen Toub in an email some time ago:

> [Toub] Whenever you use ContinueWith, I recommend being explicit about the TaskScheduler you want to target.  If you don’t specify one, it’ll end up being TaskScheduler.Current, which if you’re inside of a Task will be whatever scheduler is associated with that Task, and that means you need to test more cases.  If you instead specify TaskScheduler.Default (which targets the ThreadPool), in particular for your own internal continuations like the one in StartAsGenericTask, that’ll limit the variability in behaviors you need to test for.
[jdom] I’m using TaskContinuationOptions.ExecuteSynchronously because we expect very short continuations. Should I still use the Default scheduler?
[Toub] That’s fine.  The default scheduler will always allow inlining of continuations, meaning that with ExecuteSynchronously, TPL will be able to execute the continuation synchronously on the thread completing the antecedent without needing to queue to some other thread.