@gabikliot 2015-12-10
Got you about canceled task!

Re: `TaskScheduler.Default` - so in case of `ExecuteSynchronously`, it will inline on the current thread but switch the TaskScheduler from Current (X in my example) to Default (thread pool)?