@gabikliot 2015-12-10
This is OK to merge from my perspective. Looks good.
Not sure about the timing of 1.2.0 merge, so not going to merge until 1.1.0 is out.