@sergeybykov 2015-12-10
1.1.0 is already in a branch. We are not holding any PRs that are ready.