@jdom 2015-12-10
thanks for merging. 
BTW, for the record @gabikliot , I'm not sure what you mean by "switch the TaskScheduler from Current to Default", the task scheduler only comes in before invoking/inlining the continuation, but because the `Ignore` method doesn't return a task, this ThreadPool scheduler is only used to execute the continuation inline and there is no "permanent" switch or anything like that. People can still await the original task by using the Current scheduler and so on if they want (nevertheless Ignore is typically used when you don't care to continue it, but still)