@gabikliot 2015-12-10
@jdom, I was indeed talking about  the ThreadPool scheduler that is only used to execute the continuation inline (of course there is no "permanent" switch or anything like that). The current thread executes the Ignore method of TaskScheduler X. Now this thread needs to execute CW inline, but it needs to change the Current scheduler from X to ThreadPool, just for this CW. Don't know how expensive or not it is (maybe just one line, maybe more), but looks redundant. That is what I was asking about: " it will inline on the current thread but switch the TaskScheduler from Current (X in my example) to Default (thread pool)"