@jdom 2015-12-10
I think I understand what you are asking now. Well there is not really a "switch", as I understand it, each Task is scheduled on a scheduler, so if we were to use Current, it would be at least exactly the same overhead as scheduling on Default. Nevertheless we know Default always inlines and is very optimized for this particular scenario. We cannot know that about Current (although most of the times it will be the OrleansTaskScheduler, and I haven't checked how it handles this situation), but it can't be better than the heavily tuned default scheduler, either if it is the OrleansTaskScheduler or a custom one.