@gabikliot 2015-12-10
Great points about inlining in our scheduler.
Unrelated to that PR you might want to take a look at how we currently do inlining:

Per activation scheduler: https://github.com/dotnet/orleans/blob/master/src/OrleansRuntime/Scheduler/ActivationTaskScheduler.cs#L105
Global scheduler: 
https://github.com/dotnet/orleans/blob/master/src/OrleansRuntime/Scheduler/OrleansTaskScheduler.cs#L309

You might find out those can be optimized further to do more inlining, or do the checks more efficiently. I can tell you for sure that getting the most optimized inlining was not on my radar when implementing those, it was mostly to get it right and inline only what is allowed to be inlined without breaking the single threaded execution semantics. So there might be opportunities for optimizations. I can provide more details on how the scheduler works.