@jdom 2015-12-10
ok, I'm not that familiar to optimize them right now, but it something we might consider in the future.
But for sure it shows that in this particular implementation at least (the `Ignore()` extension method), we certainly don't need the Orleans context or anything like that. I almost kept the default CW overload (which is to use Current) until I remembered Toub's conversation. I'm more glad now that I did the change. The benchmarks I had were outside the Orleans scheduler entirely, so I would expect even less CPU overhead now :)